+++ pygeos/coordinates.py
@@ -4,19 +4,26 @@ import numpy as np
 __all__ = ["apply", "count_coordinates", "get_coordinates", "set_coordinates"]
 
 
-def apply(geometry, transformation):
+def apply(geometry, transformation, include_z=False):
     """Returns a copy of a geometry array with a function applied to its
     coordinates.
 
-    All returned geometries will be two-dimensional; the third dimension will
-    be discarded, if present.
+    With the default of ``include_z=False``, all returned geometries will be
+    two-dimensional; the third dimension will be discarded, if present.
+    When specifying ``include_z=True``, the returned geometries preserve
+    the dimensionality of the respective input geometries.
 
     Parameters
     ----------
     geometry : Geometry or array_like
     transformation : function
-        A function that transforms a (N, 2) ndarray of float64 to another
-        (N, 2) ndarray of float64.
+        A function that transforms a (N, 2) or (N, 3) ndarray of float64 to
+        another (N, 2) or (N, 3) ndarray of float64.
+    include_z : bool, default False
+        Whether to include the third dimension in the coordinates array
+        that is passed to the `transformation` function. If True, and a
+        geometry has no third dimension, the z-coordinates passed to the
+        function will be NaN.
 
     Examples
     --------
@@ -28,9 +35,16 @@ def apply(geometry, transformation):
     True
     >>> apply([Geometry("POINT (0 0)"), None], lambda x: x).tolist()
     [<pygeos.Geometry POINT (0 0)>, None]
+
+    By default, the third dimension is ignored:
+
+    >>> apply(Geometry("POINT Z (0 0 0)"), lambda x: x + 1)
+    <pygeos.Geometry POINT (1 1)>
+    >>> apply(Geometry("POINT Z (0 0 0)"), lambda x: x + 1, include_z=True)
+    <pygeos.Geometry POINT Z (1 1 1)>
     """
     geometry_arr = np.array(geometry, dtype=np.object)  # makes a copy
-    coordinates = lib.get_coordinates(geometry_arr, False)
+    coordinates = lib.get_coordinates(geometry_arr, include_z)
     new_coordinates = transformation(coordinates)
     # check the array to yield understandable error messages
     if not isinstance(new_coordinates, np.ndarray):
@@ -110,8 +124,10 @@ def get_coordinates(geometry, include_z=False):
 def set_coordinates(geometry, coordinates):
     """Returns a copy of a geometry array with different coordinates.
 
-    All returned geometries will be two-dimensional; the third dimension will
-    be discarded, if present.
+    If the coordinates array has shape (N, 2), all returned geometries
+    will be two-dimensional, and the third dimension will be discarded,
+    if present. If the coordinates array has shape (N, 3), the returned
+    geometries preserve the dimensionality of the input geometries.
 
     Parameters
     ----------
@@ -126,10 +142,27 @@ def set_coordinates(geometry, coordinates):
     [<pygeos.Geometry POINT (1 2)>, <pygeos.Geometry LINESTRING (3 4, 5 6)>]
     >>> set_coordinates([None, Geometry("POINT (0 0)")], [[1, 2]]).tolist()
     [None, <pygeos.Geometry POINT (1 2)>]
+
+    Third dimension of input geometry is discarded if coordinates array does
+    not include one:
+
+    >>> set_coordinates(Geometry("POINT Z (0 0 0)"), [[1, 1]])
+    <pygeos.Geometry POINT (1 1)>
+    >>> set_coordinates(Geometry("POINT Z (0 0 0)"), [[1, 1, 1]])
+    <pygeos.Geometry POINT Z (1 1 1)>
     """
     geometry_arr = np.asarray(geometry, dtype=np.object)
     coordinates = np.atleast_2d(np.asarray(coordinates)).astype(np.float64)
-    if coordinates.shape != (lib.count_coordinates(geometry_arr), 2):
+    n_coords = lib.count_coordinates(geometry_arr)
+    if coordinates.ndim != 2:
+        raise ValueError(
+            "The coordinate array should have dimension of 2 "
+            "(has {})".format(coordinates.ndim)
+        )
+    if (
+        (coordinates.shape[0] != lib.count_coordinates(geometry_arr))
+        or (coordinates.shape[1] not in {2, 3})
+    ):
         raise ValueError(
             "The coordinate array has an invalid shape {}".format(coordinates.shape)
         )